Altera script para não gerar o csv de apelidos caso já exista

+++ gera_entrada_google_trends.R
@@ -36,16 +36,19 @@ print(args)
 proposicoes_filepath <- args$proposicoes_filepath
 apelidos_filepath <- args$apelidos_filepath
 
-proposicao <- read_csv(proposicoes_filepath)
-
-df_google_trends <- 
-  proposicao %>% 
-  mutate(apresentacao = as.Date(ymd_hms(data_apresentacao)),
-         nome_formal = paste0(sigla_tipo, " ", numero, "/", year(apresentacao)),
-         apelido = iconv(apelido, from="UTF-8", to="ASCII//TRANSLIT")) %>% 
-  select(id_leggo, apelido, nome_formal, apresentacao, id_ext, casa) %>% 
-  group_by(apelido) %>% 
-  arrange(apelido, desc(apresentacao)) %>%  
-  slice(n())
-
-write_csv(df_google_trends, apelidos_filepath)
+## Read PLs list
+if (!file.exists(apelidos_filepath)) {
+  proposicao <- read_csv(proposicoes_filepath)
+  
+  df_google_trends <- 
+    proposicao %>% 
+    mutate(apresentacao = as.Date(ymd_hms(data_apresentacao)),
+           nome_formal = paste0(sigla_tipo, " ", numero, "/", year(apresentacao)),
+           apelido = iconv(apelido, from="UTF-8", to="ASCII//TRANSLIT")) %>% 
+    select(id_leggo, apelido, nome_formal, apresentacao, id_ext, casa) %>% 
+    group_by(apelido) %>% 
+    arrange(apelido, desc(apresentacao)) %>%  
+    slice(n())
+  
+  write_csv(df_google_trends, apelidos_filepath)
+}
